[FIX] 디폴트 이미지 삭제 안되게끔 수정

+++ src/main/java/com/nada/server/commons/S3Utils.java
@@ -40,6 +40,7 @@ public class S3Utils {
 
     public void deleteBackground(String url){
         String key = url.split("https://nada-server.s3.ap-northeast-2.amazonaws.com/")[1];
+        if(key.contains("default/")) return;
         DeleteObjectRequest deleteObjectRequest = new DeleteObjectRequest(bucket, key);
         amazonS3Client.deleteObject(deleteObjectRequest);
     }
